Added other shit

+++ skeleton/src/uk/ac/ed/bikerental/BankDetails.java
@@ -5,10 +5,10 @@ import java.util.Objects;
 import java.time.LocalDate;
 public class BankDetails{
 	private String name;
-	private BigDecimal credit_card_nr;
+	private int credit_card_nr;
 	private LocalDate expiry_date;
-	private BigDecimal security_code;
-	public BankDetails(BigDecimal credit_nr,String Name, LocalDate expiry,BigDecimal code){
+	private int security_code;
+	public BankDetails(int credit_nr,String Name, LocalDate expiry,int code){
 		this.name = Name;
 		this.credit_card_nr = credit_nr;
 		this.expiry_date = expiry;
@@ -17,13 +17,13 @@ public class BankDetails{
 	public String getName(){
 		return this.name;
 	}
-	public BigDecimal getCreditCardNr(){
+	public int getCreditCardNr(){
 		return this.credit_card_nr;
 	}
 	public LocalDate getExpiryDate(){
 		return this.expiry_date;
 	}
-	public BigDecimal getSecurityCode(){
+	public int getSecurityCode(){
 		return this.security_code;
 	}
